build: centos7: gcc4.8: "hack" -Wenum-compare warning
Fix warning, which appears when using gcc 4.8:

rm/rm_foms.c|450 col 18| error: comparison between 'enum m0_rm_incoming_type' and 'enum fop_request_type' [-Werror=enum-compare]
      M0_ASSERT(type != FRT_REVOKE);
                     ^

+++ lib/ut/varr.c
@@ -69,7 +69,7 @@ enum misc_params {
 };
 
 M0_BASSERT(BUFF_SIZE == M0_BITS(BUFF_SHIFT));
-M0_BASSERT(M0_0VEC_ALIGN > BUFF_SIZE);
+M0_BASSERT((int)M0_0VEC_ALIGN > (int)BUFF_SIZE);
 M0_BASSERT(!(M0_0VEC_ALIGN & (M0_0VEC_ALIGN - 1)));
 
 /* sizeof struct po2 is an integer power of two. */
